Crate foil

FOIL Flat Optimization Intermediate Language

Re-exports

pub use circuit::Relation;

pub use congruence::Graph;

pub use congruence::Id;

pub use congruence::LabelTrait;

pub use congruence::MetaData;

pub use congruence::Vert;

Modules

circuit

This module provides a general mechanism for synthesizing a circuit for a minimized Foil instance.

coil

congruence

Congruence Closure

constructors

Congruence Closure extended with constructors and projectors. The constructors module is just an example of how Foil can be used. The enum-based implementation of Fun is not necessary, just a somewhat simple way to make a small, self-contained example.

Structs

Foil

FoilConfig

Func

MappedFoil

Meta

Schema

Var

Enums

Label

Traits

MetaMapper